Split Test.DejaFu.SCT.Internal up
