+++ dejafu/dejafu.cabal
@@ -51,7 +51,8 @@ library
                      , Test.DejaFu.Conc.Internal.STM
                      , Test.DejaFu.Conc.Internal.Threading
                      , Test.DejaFu.Internal
-                     , Test.DejaFu.SCT.Internal
+                     , Test.DejaFu.SCT.Internal.DPOR
+                     , Test.DejaFu.SCT.Internal.Weighted
 
   -- other-modules:       
   -- other-extensions:
